backend_arch-main/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Utils
python-dotenv==1.0.1

# HTTP client
httpx==0.27.0
